Сделал так, чтобы при нажатии на статью переходило на страницу с более подробной информации

+++ admin-panel-service/src/main/java/com/example/admin_panel_service/dto/RequestDtoRedBookEntry.java
@@ -29,6 +29,6 @@ public class RequestDtoRedBookEntry {
     private String authors;
 
     private RequestDtoImage image;
-    private Integer view;//data in base64
+    private Long view;//data in base64
 
 }
